UX improved, bug in Timers fixed.
Tasks are always editable in place. New Task adds an empty editable task. Form for creating tasks is removed.

Switching to old Redux Saga 0.16.0 and following the code form the book solved the problem that when a task was moved to In Progress it was "stealing" the increments of other In Progress tasks.